added ability to compile updater
